pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genraitor"
version = "0.1.1"
readme = "README.md"
keywords = [
	"omics",
	"disease prediction",
	"DeepIMV",
	"Llama3",
	"metabolomics",
	"data harmonization",
	"RAG",
	"RAFT",
	"pathogens",
	"macromolecules",
]
description = "Predicts most important features for disease prediction using omics sampling and Llama3 fine-tuning"
authors = [{ name = "Matt Jensen", email = "matt.jensen@pnnl.gov" }]
dependencies = [
	"setuptools >= 70.2.0",
	"click >= 8.1.7",
	"structlog >= 24.2.0",
	"python-dotenv >= 1.0.1",
	"python-configuration >= 0.11.0",
	"torch >= 2.4.0",
	"nvidia-cudnn-cu12 >= 8.0.0.0",
	"trl >=0.9.4",
	"bitsandbytes >= 0.42.0",
	"flash-attn >= 2.6.3",
	"transformers >= 4.43.1",
	"peft >= 0.11.1",
	"numpy <= 2.0.0",
	"datasets >= 2.20.0",
	"ollama >= 0.3.0",
	"llama-index >= 0.10.52",
	"llama-index-llms-ollama >= 0.2.2",
	"llama-index-embeddings-ollama >= 0.1.2",
	"llama-index-embeddings-huggingface>=0.3.1",
	"notebook >= 7.2.1",
	"duckdb >= 1.0.0",
	"biopython >= 0.84",
	"tabulate >= 0.9.0",
	"metapub >= 0.5.12",
	"ruff>=0.6.1",
	"alignscore",
]

[tool.setuptools]
packages = ["genraitor"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_subclassing_any = true
disallow_any_decorated = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_return_any = true
exclude = ['tests']

[tool.ruff]
line-length = 88
extend-select = ["I"]
exclude = [
	"__pycache__",
	".git",
	"tests",
	"docs"
]

[tool.ruff.lint]
pydocstyle.convention = "google"
ignore = ['D203', 'D212', "E203", "E501"]
select = [
	'C',
	'F',
	'E',
	'W',
	'I',
	'N',
	'D',
	'B',
	'A',
	'COM',
	'C4',
	'T20',
	'Q',
	'SIM',
]

[tool.uv.sources]
alignscore = { path = "lib/AlignScore" }
